+++ test/media_types/loose_test.rb
@@ -3,6 +3,7 @@ require 'test_helper'
 require 'active_support/callbacks'
 require 'abstract_controller/callbacks'
 require 'abstract_controller/rendering'
+require 'rack/utils'
 require 'rack/response'
 require 'action_dispatch/http/content_security_policy'
 require 'action_controller'
@@ -16,6 +17,8 @@ require 'action_dispatch/http/response'
 require 'media_types'
 
 require 'oj'
+require 'awesome_print'
+
 
 class MediaTypes::LooseTest < Minitest::Test
   class MyResourceMediaType
@@ -71,12 +74,42 @@ class MediaTypes::LooseTest < Minitest::Test
     end
   end
 
-  def setup
-    @controller = FakeStrictController.new
-    @response = ActionDispatch::Response.new
+  class FakeLooseController < BaseController
+    allow_input_serializer(MyResourceSerializer)
+    allow_output_serializer(MyResourceSerializer)
+    freeze_io!
+
+    def action
+      data = deserialize!(request)
+      data[:_links][:self] = { href: 'https://example.org/loose' }
+
+      render_media(data, status: 201)
+    end
   end
 
-  def test_input_validates
+  def setup
+    @strict_controller = FakeStrictController.new
+  end
+
+  def test_strict_on_input
+    response = ActionDispatch::Response.new
+    loose_controller = FakeLooseController.new
+
+    content_type = 'application/vnd.mydomain.my_resource.v1+json'
+    request = ActionDispatch::Request.new({
+      Rack::RACK_INPUT => StringIO.new({ test: 1, _links: {  } }.to_json),
+      'CONTENT_TYPE' => content_type,
+      'HTTP_ACCEPT' => "#{content_type}, application/problem+json; q=0.1"
+    })
+
+    # Input is not valid because example is unexpected
+    status, _headers, _body = loose_controller.dispatch(:action, request, response)
+    assert_equal 400, status
+  end
+
+  def test_loose_on_input_only
+    response = ActionDispatch::Response.new
+
     content_type = 'application/vnd.mydomain.my_resource.v1+json'
     request = ActionDispatch::Request.new({
       Rack::RACK_INPUT => StringIO.new({ test: 1, _links: { example: { href: "https://example.org/" } } }.to_json),
@@ -86,8 +119,24 @@ class MediaTypes::LooseTest < Minitest::Test
 
     # Passing input to output should fail (output isn't loose)
     assert_raises MediaTypes::Serialization::OutputValidationFailedError do
-      @controller.dispatch(:action, request, @response)
+      @strict_controller.dispatch(:action, request, response)
     end
   end
-end
 
+
+  def test_invariant
+    response = ActionDispatch::Response.new
+    loose_controller = FakeLooseController.new
+
+    content_type = 'application/vnd.mydomain.my_resource.v1+json'
+    request = ActionDispatch::Request.new({
+      Rack::RACK_INPUT => StringIO.new({ test: 1, _links: { example: { href: "https://example.org/" } } }.to_json),
+      'CONTENT_TYPE' => content_type,
+      'HTTP_ACCEPT' => "#{content_type}, application/problem+json; q=0.1"
+    })
+
+    # Passing input to output should pass (loose link is added)
+    status, _headers, _body = loose_controller.dispatch(:action, request, response)
+    assert_equal 201, status
+  end
+end