+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 2.0.4
 
+- ✨ Add more loose tests
 - 🐛 Fix `locals` passing for HTML serialization
+- 🐛 Fix for status passing in rack-like environments
 
 ## 2.0.3